+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 3.4.1
+## 3.5.0
 
 Fixed